+++ src/Integrations/Hubspot/HubspotClient.php
@@ -178,12 +178,11 @@ class HubspotClient implements HubspotClientInterface
 			foreach ($items as $item) {
 				$name = $item['name'] ?? '';
 				$hidden = $item['hidden'] ?? false;
-				$readOnlyValue = $item['readOnlyValue'] ?? false;
 				$formField = $item['formField'] ?? false;
-				$deleted = $item['deleted'] ?: false; // phpcs:ignore WordPress.PHP.DisallowShortTernary.Found
+				$archived = $item['archived'] ?? false;
 				$fieldType = $item['fieldType'] ?? '';
 
-				if (!$name || $hidden || $readOnlyValue || !$formField || $deleted) {
+				if (!$name || $hidden || !$formField || $archived) {
 					continue;
 				}
 
@@ -269,7 +268,7 @@ class HubspotClient implements HubspotClientInterface
 			$paramsFiles
 		);
 
-		$url = $this->getBaseUrl("submissions/v3/integration/secure/submit/{$baseId}/{$submitId}");
+		$url = "https://api.hsforms.com/submissions/v3/integration/secure/submit/{$baseId}/{$submitId}"; // Not using getBaseUrl method because of different url only for this legacy endpoint.
 
 		$response = \wp_remote_post(
 			$url,
@@ -329,10 +328,7 @@ class HubspotClient implements HubspotClientInterface
 					continue;
 				}
 
-				$properties[] = [
-					'property' => $key,
-					'value' => $value,
-				];
+				$properties[$key] = $value;
 			}
 		}
 
@@ -340,11 +336,12 @@ class HubspotClient implements HubspotClientInterface
 			'properties' => $properties,
 		];
 
-		$url = $this->getBaseUrl("contacts/v1/contact/createOrUpdate/email/{$email}", true);
+		$url = $this->getBaseUrl("crm/v3/objects/contacts/{$email}?idProperty=email");
 
 		$response = \wp_remote_post(
 			$url,
 			[
+				'method' => 'PATCH',
 				'headers' => $this->getHeaders(),
 				'body' => \wp_json_encode($body),
 			]
@@ -390,8 +387,12 @@ class HubspotClient implements HubspotClientInterface
 			$folder = self::HUBSPOT_FILEMANAGER_DEFAULT_FOLDER_KEY;
 		}
 
+		$fileName = \pathinfo($file, \PATHINFO_FILENAME);
+		$fileExtension = \pathinfo($file, \PATHINFO_EXTENSION);
+
 		$options = [
 			'folderPath' => '/' . $folder,
+			'fileName' => "{$fileName}.{$fileExtension}",
 			'options' => \wp_json_encode([
 				"access" => "PUBLIC_NOT_INDEXABLE",
 				"overwrite" => false,
@@ -414,7 +415,7 @@ class HubspotClient implements HubspotClientInterface
 		\curl_setopt_array( // phpcs:ignore WordPress.WP.AlternativeFunctions.curl_curl_setopt_array
 			$curl,
 			[
-				\CURLOPT_URL => $this->getBaseUrl("filemanager/api/v3/files/upload", true),
+				\CURLOPT_URL => $this->getBaseUrl("files/v3/files"),
 				\CURLOPT_FAILONERROR => true,
 				\CURLOPT_POST => true,
 				\CURLOPT_RETURNTRANSFER => true,
@@ -430,7 +431,7 @@ class HubspotClient implements HubspotClientInterface
 		if ($code >= UtilsConfig::API_RESPONSE_CODE_SUCCESS && $code <= UtilsConfig::API_RESPONSE_CODE_SUCCESS_RANGE) {
 			$response = \json_decode((string) $response, true);
 
-			return $response['objects'][0]['url'] ?? '';
+			return $response['url'] ?? '';
 		}
 
 		return '';
@@ -551,7 +552,7 @@ class HubspotClient implements HubspotClientInterface
 	 */
 	private function getHubspotContactProperties()
 	{
-		$url = $this->getBaseUrl('properties/v1/contacts/properties', true);
+		$url = $this->getBaseUrl('crm/v3/properties/contacts');
 
 		$response = \wp_remote_get(
 			$url,
@@ -574,7 +575,7 @@ class HubspotClient implements HubspotClientInterface
 
 		// On success return output.
 		if ($code >= UtilsConfig::API_RESPONSE_CODE_SUCCESS && $code <= UtilsConfig::API_RESPONSE_CODE_SUCCESS_RANGE) {
-			return $body ?? [];
+			return $body['results'] ?? [];
 		}
 
 		return [];
@@ -587,7 +588,7 @@ class HubspotClient implements HubspotClientInterface
 	 */
 	public function getTestApi(): array
 	{
-		$url = $this->getBaseUrl('forms/v2/forms', true);
+		$url = $this->getBaseUrl('forms/v2/forms');
 
 		$response = \wp_remote_get(
 			$url,
@@ -913,18 +914,11 @@ class HubspotClient implements HubspotClientInterface
 	 * Return HubSpot base url.
 	 *
 	 * @param string $path Path to append.
-	 * @param bool $legacy If legacy use different url.
 	 *
 	 * @return string
 	 */
-	private function getBaseUrl(string $path, bool $legacy = false): string
+	private function getBaseUrl(string $path): string
 	{
-		$url = 'https://api.hsforms.com';
-
-		if ($legacy) {
-			$url = 'https://api.hubapi.com';
-		}
-
-		return "{$url}/{$path}";
+		return "https://api.hubapi.com/{$path}";
 	}
 }